feat(Segmented control): organizes component structure
